move EditMessageForm into MessageInput folder becasue it is just a MessageInput presentation component
