--- src/components/EditMessageForm/index.js
@@ -1,1 +1,0 @@
-export { default as EditMessageForm } from './EditMessageForm';
