+++ src/components/Message/MessageSimple.js
@@ -14,8 +14,7 @@ import { Attachment as DefaultAttachment } from '../Attachment';
 import { Avatar } from '../Avatar';
 import { Gallery } from '../Gallery';
 import { Modal } from '../Modal';
-import { MessageInput } from '../MessageInput';
-import { EditMessageForm } from '../EditMessageForm';
+import { MessageInput, EditMessageForm } from '../MessageInput';
 import { MessageActionsBox } from '../MessageActions';
 import { Tooltip } from '../Tooltip';
 import { LoadingIndicator } from '../Loading';
