+++ src/components/Message/MessageTeam.js
@@ -13,8 +13,7 @@ import { withTranslationContext } from '../../context';
 import { Attachment as DefaultAttachment } from '../Attachment';
 import { Avatar } from '../Avatar';
 import { Gallery } from '../Gallery';
-import { MessageInput } from '../MessageInput';
-import { EditMessageForm } from '../EditMessageForm';
+import { MessageInput, EditMessageForm } from '../MessageInput';
 import { MessageActionsBox } from '../MessageActions';
 import { Tooltip } from '../Tooltip';
 import { LoadingIndicator } from '../Loading';
