+++ src/components/Message/__tests__/MessageSimple.test.js
@@ -11,9 +11,11 @@ import Message from '../Message';
 import MessageSimple from '../MessageSimple';
 import { Modal as ModalMock } from '../../Modal';
 import { Avatar as AvatarMock } from '../../Avatar';
-import { MessageInput as MessageInputMock } from '../../MessageInput';
+import {
+  MessageInput as MessageInputMock,
+  EditMessageForm,
+} from '../../MessageInput';
 import { MessageActionsBox as MessageActionsBoxMock } from '../../MessageActions';
-import { EditMessageForm } from '../../EditMessageForm';
 
 const alice = generateUser();
 const bob = generateUser({ name: 'bob', image: 'bob-avatar.jpg' });
