+++ src/components/MessageInput/index.js
@@ -2,4 +2,5 @@ export { default as MessageInput } from './MessageInput';
 export { default as MessageInputFlat } from './MessageInputFlat';
 export { default as MessageInputLarge } from './MessageInputLarge';
 export { default as MessageInputSmall } from './MessageInputSmall';
+export { default as EditMessageForm } from './EditMessageForm';
 export { default as SendButton } from './SendButton';
